requirements.txt
pandas
mne-python
matplotlib
numpy
seaborn